Show full email template on single template page
There’s no point in collapsing […] the email on the single template
page because it’s the only thing on the page.

Users will probably be going back and forth quite a bit to edit their
template and see the changes; they shouldn’t have to expanded it evey
time to check they’ve got the heading syntax right, for example.