Split out ‘created’ as it’s own state
The sooner we can surface notifications in the admin app, the more
responsive it feels. Previously we were ignoring ‘created’
notifications, which is the very earliest state.

It needs to be it’s own ‘tab’ on the jobs page because it’s a different
thing from ‘sending’ (which happens much slower in reality, with a long
tail). And ‘queued’ is probably a more human word for it than ‘created’.

This commit also removes the ‘processed’ state because that’s a very us
thing, and has a confusing partial overlap with that other states.

So the four tabs are now:
```
Created | Sending | Delivered | Failed
```