Remove check for statistics not being on job
A job will always have statistics.

It’s always assigned:
https://github.com/alphagov/notifications-api/blob/bcfa83de79e11856f12a7581b6519ec4a66dad36/app/job/rest.py#L48

And the structure should always be the same, even if the counts are
zero because they’re generated from this query:
https://github.com/alphagov/notifications-api/blob/668e6c9716cfffdb987cbf8abe20c14576864a1c/app/dao/jobs_dao.py#L11-L22

This line suggests that it’s a hangover from the aggregate tables:
https://github.com/alphagov/notifications-admin/commit/8c159da3eaf3a8100bdc2603b4addf25b9fe573d#diff-9886486bf41b0680d23588b190c252eaL24

Since it’s no long necessary this commit removes it.